Revert "edit: update another float"
This reverts commit 7e324b33b7e7b2dfbb5e26fbbe140ba4c182b4d3.

+++ UnityWorkspace/Assets/MW_Scripts/Networking/SerializeGame.cs
@@ -53,7 +53,7 @@ public class SerializeGame //: MonoBehaviour //uncomment when testing
 
 					//make new tile
 					LandType lType = (LandType)Enum.Parse (typeof(LandType), landType, true);
-					Tile myTile = new Tile (lType, myGameBoard, Convert.ToSingle (boardPositionX), Convert.ToSingle (boardPositionY));
+					Tile myTile = new Tile (lType, myGameBoard, Convert.ToInt32 (boardPositionX), Convert.ToInt32 (boardPositionY));
 					Vector2 gamePos = new Vector2 (Convert.ToSingle (gamePositionX), Convert.ToSingle (gamePositionY)); //needs to be a float aka "single"
 					myTile.gamePosition = gamePos;
